Expose ServerSocket to subclasses of OioServerSocketChannel.
Motivation:
Currently when trying to subclass OioServerSocketChannel and override
doReadMessages there is no way to access the socket, hence no way to
add a different SocketChannel implementation.

Modifications:
Add a protected serverSocket() method that returns the ServerSocket
similar to what AbstractNioChannel does with its javaChannel() method.

Result:
It is possible to subclass OioServerSocketChannel and have a subslass
return a different SocketChannel implementation.