Return only the requested allowed headers instead of all.
Motivation:
Currently the CorsHandler will return all of the allowed request headers
configured. But a client might only send a subset of these headers in a
CORS request and should only have those headers included in the
response.

Modifications:
Enabled CorsHandler to return only the allowed request headers that
exist in the actual request and not simply all.

Result:
The response will include only the request headers that the client
specified and not all of the headers that the server allows. This can
make testing simpler and also reduces the amount of data sent.