Adding sesssionId() method to SessionState
Motivation:
SockJS handlers may need access to the sessionId for various reasons.
One might be to track state for a certain session.

Modifications:
Added a sessionId method to the SessionState interface and also added an
AttributeKey which is available to SockJS handlers.

Result:
SockJS handlers can get access to the SockJS session identifier.