Adding support for echoing the request origin for CORS.
Motivation:
When CORS has been configured to allow any origin, and at the same time
is allowing credentials, this causes an error from the browser because
the wildcard origin, '*', failing the CORS request.
When the response 'Access-Control-Allow-Credentials' is true the
'Access-Control-Allow-Origin' must be an actual origin.

Modifications:
Changed CorsHandler setOrigin method to check for the combination of any
origin and allowCredentials and if the check matches echo the CORS
request 'Origin' value.

Result:
This addition enables the echoing of the request 'Origin' value as the
'Access-Control-Allow-Origin' value when the server has been configured
to allow any origin in combination with allowCredentials.
This allows client requests to succeed when expecting the server to
response with any origin and at the same time be able to send cookies
by setting xhr.withCredentials=true.
A concrete example of this is the SockJS protocol which expects
behaviour.

Conflicts:
	codec-http/src/main/java/io/netty/handler/codec/http/cors/CorsHandler.java
	codec-http/src/test/java/io/netty/handler/codec/http/cors/CorsHandlerTest.java